tilp: Replace GLADE_CB macro with TILP_EXPORT
The macro was already custom as it needed to be modified for our use
of visibility=hidden. The GLADE_CB macro was defined in an otherwise
libglade specific header so the new tilp specific macro was placed
in tilp_defs.h so that those glade specific files can be removed in a
later commit.

Signed-off-by: Jon Sturm <jonimoose@gmail.com>