tilp: Remove remaining glade references and rework create_pixbuf
Now that all the interfaces have been converted to gtkbuilder
we can remove all of the glade related files. This includes
the glade generated support.c and support.h but we still
used create_pixbuf from them so rework it and move it to
tilp_paths

TODO: in the future we should move the app icon to the proper
fd.o specified folder which would simplify loading it on *nix.

Signed-off-by: Jon Sturm <jonimoose@gmail.com>